Don't use swift plan for existing passwords
We'll use the existing stack environment instead.
When we move to ephemeral heat stack, we'll use
a file. Rotate passwords would generate an environment
with passwords which can then be used in subsequent
deployments.

Also removes the legacy heat resource passwords as we
probably allow ffw upgrades from queens to
wallaby and these passwords would already be in stack
environment.

Depends-On: https://review.opendev.org/c/openstack/python-tripleoclient/+/765808
Change-Id: I7e081a1831bc00e91cd6967c4c404c0037c85d17