Removing tempest container build
The tempest container was used with validate-tempest, which is right now
deprecated, and all the jobs running upstream are using os_tempest,
which doesn't rely on containerized tempest, but only on the rpm
packages and/or install tempest directly from git.
The tempest container image is no longer being used.

Depends-On: https://review.opendev.org/c/openstack/tripleo-quickstart/+/778447

Related-Bug: #1916875

Change-Id: I4dc869019a04bf092f8c219e470be55f3bc682d3