Reduce fact gathering
This change updates the default subset to only be the minimum list of
facts for ansible.  In the deployment we need to make sure that we only
collect the minimum amount of facts to perform the tasks at hand rather
than assume we have all the facts.

Depends-On: https://review.opendev.org/c/openstack/tripleo-heat-templates/+/777002
Depends-On: https://review.opendev.org/c/openstack/tripleo-ansible/+/777004
Related-Bug: #1915761
Change-Id: I61affe32dfc9eacbe514893168397c9b291b8713
(cherry picked from commit 0e81eb420e8974d7815b98e084a94885a1ff1b8b)