Add an Ansible execution environment to our Build
This change adds an ansible execution environment to our build tools,
which will allow us to run with execution environments via tripleoclient
in the future. This build simply pulls in the base ansible-runner
image, and provides an interface to download/install roles and
collections as needed.

It also contains:
- the tripleo-ansible roles and collections contents via RPM.
- tripleo-ansible ansible contents are symlinked back with
  the /runner directory.
- ansible-runner settings included in /runner/env
- custom entrypoint for tripleo specific usecase, which later
  on appends the ansible-runner entrypoint

This image will be consumed in
tripleo-ci-centos-9-singlenode-external-compute-via-ee job defined in
https://review.opendev.org/c/openstack/tripleo-ci/+/843836.

Depends-On: https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/852720

Change-Id: I56cabb9aa8483c5039b5920db96a00509dc730cc
Signed-off-by: Kevin Carter <kecarter@redhat.com>
Co-authored-by: Chandan Kumar <chkumar@redhat.com>