Remove layout override and wire up buildimage template instead
This wires up the tripleo-buildimage-jobs and updates buildimage
job overrides in the check layout. The depends-on adds the
files: matches in the template layout so no need to override here.
The gate layout overrides are removed altogether as we don't want
the dependencies: optimization in the gate. Part of wider tripleo-ci
optimization [1].

[1] https://review.opendev.org/q/topic:tripleo-ci-reduce
Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/777083

Change-Id: I24615a067a56ac226b07c2ebcf477b3682dfed2f