Removes unneeded zuul layout overrides - templates for consistency
As part of [1] this removes the unneeded zuul layout overrides.
All the jobs removed here are already included with the relevant
templates. The files: matches that were used for the scenario
standalones are moved to the template layout via the depends-on.

The build-containers-ubi8 job is not needed as it is exercised
by the content providers and we have removed this job from other
tripleo layouts.

[1] https://review.opendev.org/q/topic:tripleo-ci-reduce
Depends-On: I633bc01bc1a8cd8d1216e503eaa9a43aef9a7fa0

Change-Id: I95efa61361882875505e5e359dece23a5a67bfc9
(cherry picked from commit 22e5366de13a342ca6f65173ce069513fa80e1f5)