[stable-only] Fix legacy password handling
We pick some of heat legacy passwords (from heat resources),
while generating the passwords. However, if the plan has
been deleted from some reason and does not have the old
passwords as in the stack environment, this would result
in change of these passwords like RabbitCookie/PcsdPassword
and would lead to disruptions during deployment. This
patch fixes to only use them if they are not in the stack
environment.

This code has been removed from master after plan removal.

Closes-Bug: #1918566
Change-Id: I6b2f0ef916a950fe6f92ee4f8848bab92a7b1c85