Add passwords_env parameter to generate_passwords
Add a new parameter, passwords_env to the generate_passwords function
that can be used to pass in a Heat environment dict of existing password
values. If the value is not None, it will be used instead of attempting
to query Heat for the stack environment.

The paramter will allow for re-using a set of existing passwords without
having to have an existing Heat stack.

Signed-off-by: James Slagle <jslagle@redhat.com>
Change-Id: I99b488842261775dbbb0308c070e440f71b5be41