Use quay.ceph.io to pull the Ceph Dashboard related images
The purpose of this change is just to switch to quay.ceph.io
and pull from that location the Ceph Dashboard related images.
quay.ceph.io is maintained by the Ceph organization and can
be trusted to find the tags that we need.

This is not a clean cherry-pick since grafana is bumped to
5.4.3, which is the same version (for nautilus) used on ussuri
and victoria.

Closes-Bug: #1920873
Change-Id: If5ec5d4fa0d4d835561b5963d244e5172fe82cc8
(cherry picked from commit 72c7d4ac71d33c8803120767108c6bf2ffeb2edc)