Don't install dhcp-all-interfaces for redhat 8 onwards
This element was needed for centos-7 and earlier because by default
only the first interface had dhcp enabled. This is no longer the case
with centos-8, every interface has dhcp enabled. Additionally this
element is causing issues in some environments.

This change removes the element from all *-8 images since it no longer
serves a useful purpose and is causing problems in some cases.

Change-Id: I2cc79caeabd9d1b85f54320a83e5f36a64003303