Include both Ceph Octopus and Pacific in the content provider
I9d38c1dbadb968f8c7721c69c57002f159cea619 changed the default
Ceph container to Pacific but it took time for multiple patches
to merge so there was a period of time a job was asking for a
container which was different from what was available. This patch
makes both available in the content provider (brings Ocotpus back)
until the transition is complete and then it can be reverted. It
establishes a pattern for how to make these transitions in the
future.

Closes-Bug: #1923529
Change-Id: Ia00caf064921da362395afbfeb1b0259ff244c4a