Remove sysfsutils from nova container
sysfsutils was an optional depency of nova for
fiber channel storage integration.
sysfsutils and systool shoudl nolonger be used in nova directly
as a result this depency shoudl nolonger be required.

sysfsutils is going away in centos 9 which will be the basis OS
for new versions so we should remove it from the package list in the
nova-compute container.

direct usage of sysfsutils/systool was removed a while ago but it is
still used by fiber channel in os-brick, but os-brick does require
the package correctly.

Change-Id: I36cb49007942d95f401eda1719a6eac327479503