Skip port if missing mandatory tag
Instead of bailing out if a port does not have the
mandatory tags when building ports_by_role_and_host
simply skip the port.

Also skip ports for which the network_id is not in
'networks_by_id'. This means it is a OVN bridge MAC
port and should be ignored.

Related: blueprint network-data-v2-ports
Closes-Bug: #1928469
Depends-On: https://review.opendev.org/790687
Change-Id: Ife65326509e1d5fe7689c1baec17d902ec848c0d