Disable neutron inventory src if no ctlplane ports
ctlplane ports created by THT Heat Server
resources, or nova less without --network-ports/--network-config
enabled, does not have the 'tripleo_stack_name' tag. We
shouldn't use neutron as a source if no ctlplane ports are
tagged with the 'tripleo_stack_name'.

This patch add's a check for to ensure atleast 1 port is in the
ctlplane network. If no such port is found _get_neutron_data
returns None, and the inventory generation will continue to use
the heat stack as the source.

Also filter "tripleo_service_vip" and "tripleo_vip_net" ports
when getting the ports so that the test doesn't match on a
control plane virtual IP.

Closes-Bug: #1928926
Change-Id: Ida1cc17ae6b9adf4275e0616706158bfc3b93204