Fix site-packages in horizon's extend_start.sh
We were relying on an unset variable from kolla:
KOLLA_DISTRO_PYTHON_VERSION to figure out the
location of site-packages to bootstrap horizon
and load its various plugins. Since we only support
python3, we can drop the need for a config variable,
and rely on facts from the system to find our way.

Change-Id: I428723525a334f7bf2235d8e6ae62da6d5d00bb7
Closes-Bug: BZ#1928291
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>