Remove libcgroup-tools from libvirt container
libcgroup-tools was added to the libvirt container image
with I469b4f51c57e1651f42064b0134aa3916af72d80.

Since OCP on OCP is right now discontinued and libcgroup
is planned to be removed from RHEL 9, lets remove it also
from the libvirt container image.

If there is the plan to start it again, we could probably
use systemd-run instead.

Change-Id: Ic946d0ef686cb178e0f02db012f2767c9ec5d6d4