Remove references to tripleo-validations legacy directory
'/usr/share/openstack-tripleo-validations' doesn't exist anymore since
Train. Tripleo-validations installs its bits in '/usr/share/ansible/'.
It is now quite safe to remove the references to this legacy directory.

**Note**
This patch will be applied to master and stable/wallaby only. This logic
will be kept as is for stable/train like this directory may still be
present during an upgrade.

Change-Id: Ib96e2c0dd789be2be6bdaf66682b985cb864baf8
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>