Do not fail if initiatorname.iscsi does not exist
In iscsid container image initiatorname.iscsi file is removed as it
needs to be different on each container. Until CentOS8, this file was
created when installing iscsi-initiator-utils but in CentOS 9 preview
this file is created by a systemd service if it does not exist, so we
shouldn't assume it will always exist.

Change-Id: Iae51282a61d16edb5a67162fca42ffcca7ba5d80