Add TypeError exception when parsing addresses
When parsing addresses the value could be None, if that is the case, we
need to handle the result and return it accordingly. This change adds
the expected TypeError when dealing with a null value.

Additionally the overcloudrc will now process the data ahead of time and
ensure that the items in the array have an assosiated value before
feeding the data into the mapped function.

Change-Id: I5a1900d4e50395065b6b56876ab9ee10f01fae05
Signed-off-by: Kevin Carter <kecarter@redhat.com>
(cherry picked from commit eaeb230f838f5a806daf70845fe970c77534e1bc)