Add missing ipa images to overcloud-hardened-images-uefi
CI jobs which build images will include a config file then run the
command to build all images. This means that
overcloud-hardened-images-uefi needs to do the same thing as
overcloud-full, and have the image definition for ironic-python-agent
as well as overcloud-hardened-uefi-full.

Without this change the ironic-python-agent image won't be built and
the CI job will fail.

Blueprint: whole-disk-default
Change-Id: I8b5ad48920db4311bf6a63d61fc486446bef8c22