Check for no neutron network in inventory
The host network or SSH network specified with --overcloud-ssh-network
may not exist in Neutron. There is no requirement that it does. Add a
check and log a warning instead of trying to use the None object when
the network is not found. When it's not found, neutron data is not added
to the inventory.

Signed-off-by: James Slagle <jslagle@redhat.com>
Change-Id: I20d6ae39511e368e724522d133a133d1c864cb1b