Remove ironic-inspector dnsmasq healthcheck
This check is broken because healthcheck_port uses the user of the
process, but the /proc files accessed are owned by root. This is
probably because dnsmasq starts as root to bind to priviliged port 67,
then drops to the dnsmasq user.

This healthcheck likely doesn't add much value since it is a
self-contained tool which will be bound to the port if it is running.

See https://bugzilla.redhat.com/show_bug.cgi?id=1961237

Change-Id: Ic23e80465356f0a336621008e0a9d5b0c295d816