Update tcib for CentOS Stream 9
It moves the EL8 and EL9 packages into two conditional
and also introduces python_version for both EL versions.

https://review.opendev.org/c/openstack/python-tripleoclient/+/807098
adds the support for passing extra parameters and based on those
parameter above distro and python version conditionals
will be evaluated.

Since collectd packages and qdrouterd packages are not available
currently so are restricting to EL8.

Note: /usr/share/zoneinfo comes from tzdata and this directory
does not exists on base image. On reinstalling in a running
container, the directory exists but on a image it does not
exists. So In order to fix it, we are removing the tzdata
rpm and reinstalling it as a workaround.

Related-Bug: #1946461

Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/800577

Note: It is not a clean cherry-pick. It collects other CS9
changes https://review.opendev.org/q/%28topic:%2522cs9%2522+OR+hashtag:%2522cs9%2522%29+status:merged+project:openstack/tripleo-common
in a single patch.

Below is the list of patches included in this change:
* https://review.opendev.org/c/openstack/tripleo-common/+/816321
* https://review.opendev.org/c/openstack/tripleo-common/+/820041
* https://review.opendev.org/c/openstack/tripleo-common/+/820382
* https://review.opendev.org/c/openstack/tripleo-common/+/814119
* https://review.opendev.org/c/openstack/tripleo-common/+/813745
* https://review.opendev.org/c/openstack/tripleo-common/+/808044

Already Backported:
* https://review.opendev.org/q/I544d117e03037c8a4744139c36bc63b978c489dd
* https://review.opendev.org/q/I6c3798c40b7f5c241a7eb13d2ae01cd23fd943b5

Signed-off-by: Ronelle Landy <rlandy@redhat.com>
Co-authored-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I23c64637e161c8054332c4e4761cd82802f3ae1f
(cherry picked from commit f558eaa75a12cc2f1c3147f61c57510e51859c99)