Update tcib for CentOS Stream 9
It moves the EL8 and EL9 packages into two conditional
and also introduces python_version for both EL versions.

https://review.opendev.org/c/openstack/python-tripleoclient/+/807098
adds the support for passing extra parameters and based on those
parameter above distro and python version conditionals
will be evaluated.

Since collectd packages and qdrouterd packages are not available
currently so are restricting to EL8.

Note: /usr/share/zoneinfo comes from tzdata and this directory
does not exists on base image. On reinstalling in a running
container, the directory exists but on a image it does not
exists. So In order to fix it, we are removing the tzdata
rpm and reinstalling it as a workaround.

Related-Bug: #1946461

Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/800577

Signed-off-by: Ronelle Landy <rlandy@redhat.com>
Co-authored-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I23c64637e161c8054332c4e4761cd82802f3ae1f