Add dnsmasq to ironic-pxe container
In order for us to handle RHEL9's removal of tftp-server, we need
to migrate to another TFTP server to support PXE operations, and
luckily enough, dnsmasq has this capability.

With the package in place, we should be able to invoke it
for the container to run and then eventually remove tftp-server.

Change-Id: I28a3e1200567236d4f78bfddb7899490bbea267f