Add security_reset.expect.10.5 from kolla for mariadb>10.5
https://review.opendev.org/c/openstack/tripleo-common/+/800580 adds
the support of building containers on CentOS Stream 9.

mariadb>10.5 does not works with
/usr/share/openstack-tripleo-common-containers/container-images/kolla/mariadb/security_reset.expect
file. It needs
https://opendev.org/openstack/kolla/src/branch/master/docker/mariadb/mariadb-server/security_reset.expect.debian
to be placed with in kolla/mariadb directly so that it
will work with mariadb>10.5.

The file is renamed to security_reset.expect.10.5 to
avoid debian in the end.

Since above filepath is hardcoded in the mariadb
container so putting it in c9 container build
patch is not working.
So we are moving it to a seperate patch.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I6609345d0e60af62ca75e4bd9b3ed63167a3bc70
(cherry picked from commit e08affe816a700c996975181b84594cf02431b63)