[c9] replacing genisoimage with xorriso
In RHEL9, genisoimage was replaced with xorriso as per BZ#1971840

The reason is because genisoimage is no longuer maintained as expected.
Xorriso is a dropin replacement for genisoimage. mkisofs is symlinking
to alternatives which in turn is symlinked to either genisoimage or
xorriso.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=2002949
Change-Id: I6c3798c40b7f5c241a7eb13d2ae01cd23fd943b5