add kernel-modules-extra in overcloud image needed for HWOL
In case of ovs-hw-offload, without k-m-e installed,
it will have a performance impact because of excessive upcalls
due to the lack of act_ct.

This patch is intended to provision k-m-e in overcloud image
building.

act_ct would be added to kernel-core in rhel9.
So, k-m-e doesnt required in rhel9,centos9.

Closes-Bug: #1946227
Change-Id: If398172876059559ce70ebb868ada82da009a02d