[EL8] Add nodejs:14 module stream
nodejs is a dependency of novnc in EL8, so it gets installed in the
nova-novncproxy image. Not specifying a module stream means that the
install can default to an unsupported version of nodejs.

Explicitly enabling nodejs:14 prevents that.

The EL9 build of novnc does not require nodejs so this is not a concern
there.

Change-Id: I973dc924cfc5ba31aeb0ec19805c0f987a0221a0