Fix namespace-less containers in image-serve
The build_catalog() function expects the manifests, tags and blobs be
available before the catalog will recognized a container image. The
previous spot where this function was called was after the manifests but
before the tags are created.  This causes the newly uploaded container
not to be listed in the catalog. This change moves the build_catalog()
to last so new images get correctly added to the catalog.

Change-Id: I3f08fe0efef1e830b034fab06574a05dff89ccc0
Closes-Bug: #1950352