Stop creating manifest .htaccess
This is no longer needed as this isn't dynamic so we can moved it to the
image-server httpd configuration. Additionally MultiviewsMatch is
causing problems so we don't want that anymore.  This file will be
ignored with the updated configuration so this just stops it from being
created going forward.

Depends-On: https://review.opendev.org/c/openstack/tripleo-ansible/+/820507
Change-Id: I307be36cfdd26510bbe5b48365d6c04881f61d1f
Related-Bug: #1953198
(cherry picked from commit 6a34b52982f3d4d004f4dff30cd9ff4aeb33129b)