Add tempest to sudoers for tempest container image
Required to configure TLS in the tempest pod init when used by
osp-director-operator.

Workaround cirular dependency: tempest_sudoers will not exist until this
patch is included in the openstack-tripleo-common-containers rpm.
Fallback to using an existing sudoers file with sed, then followup with
a patch to remove this logic once the package is updated.

Depends-On: I039c49e6f5872b02c1fdc1690594e0d9bef303f5

Change-Id: I2d06904cc89e398f82ffc8b4f2d862427507d0dc