Add exception handler for node attributes
Node mac and port attributes are assumed to be strings, however, if a
mac address is not specifically quoted in YAML format, it could be
rendered as an INT. Mac address using colons for separation, could
result in an INT due to python yaml attempting to render the value.

This change adds an exception handler for the edge case and will inform
the user that the values are broken for a particular node.

Change-Id: I2105869a89877018b217aabc47c32ae44026f41d
Signed-off-by: Kevin Carter <kecarter@redhat.com>