Reinstall centos-stream-release from centos repos
In CentOS Stream 8 and 9 container image, centos-stream-release
package is already available in the container image. But in UBI,
it is not available. It can be installed from the base OS repo.

It removes the logic of downloading the centos-stream-release
rpm and installing it. As it was added earlier when we were
adding the initial support for EL9.

since redhat-release is a part of centos-stream-release
In UBI 8, redhat-release is a seperate package, so it's rpm
needs to be removed otherwise it will end up with skip broken
package.

On removing redhat-release, /etc/os-release file also gets deleted.
So we need to first remove redhat-release rpm if present and then
install centos-stream-release package.

It also removes dnf-core-plugins package installation as we donot
need to download the centos-stream-release rpm.

Closes-Bug: #1958202
Related-Bug: #1958240
Closes-Bug: #1962786

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ia06c6770ea567656b5ce380826bdb120ca32a1a8
(cherry picked from commit b09fb3fef0896e812ca9ce3e6dc74f297494203a)