Switch from grub2 to bootloader element for overcloud-full
A lot of effort has gone into the bootloader element doing the right
thing in all cases, managing the transition to BLS, and handling
legacy boot with UEFI boot.

overcloud-full and other partition images have not received the
benefit of this work because they use the much simpler grub2 element.
Now that overcloud-hardened-uefi-full is the default, overcloud-full
may become harder to support over time. Switching to the bootloader
element will reduce the support matrix for the boot handling of this
image.

This changes fixes an issue where the kernel args root option is
incorrect, because the bootloader element does a full kernel options
regeneration.

Depends-On: https://review.opendev.org/c/openstack/diskimage-builder/+/826976
Closes-Bug: #1958230
Change-Id: I9a0936c415485166962194104c7c54ee520e9516