Remove rhel specific packages for ironic-conductor
The ovirt driver which requires python-ovirt-engine-sdk4 is one of
the staging drivers, and it has been explicitly unsupported for a few
releases now, so it is removed in this change.

The python3dist\(ansible\) package is the old way of installing
ansible, and enables the ansible deploy driver. However this
driver has never been considered supported, so rather than correct the
package name, it is removed in this change.

Change-Id: I415e701b9fdd533299b343891cef920290f4d7a6