Support multiple mounts without explicit nbd devices
Currently the default nbd device is /dev/nbd0, so mounting and
unmounting one image works without specifying a -n argument.

However mounting multiple images requires an explicit correct -n
/dev/nbdX for both tripleo-mount-image and tripleo-unmount-image
calls, which gets difficult to manage.

This change fixes that by choosing an appropriate NBD device in each
case. For mount, the lsblk PTTYPE is set for devices in use, so
iterate the devices until a free one is found. For unmount, look at
the 'mount -l' output and trace the root mount path back to the NBD
device which is in use.

Change-Id: I1f221f3fb4dd70c919168d52c53e03c3b2535c04
Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/852068