[Wallaby-Only] remove remove-machine-id element
https://review.opendev.org/c/openstack/tripleo-common/+/823148
cherry-pick tries to remove element from stable/wallaby but there
are other files where this element is mentioned.

In order to remove it completely from stable/wallaby,
we need to remove the reference for the same in favor
of https://review.opendev.org/c/openstack/diskimage-builder/+/818851

https://review.opendev.org/c/openstack/tripleo-puppet-elements/+/824426
removes the element code from tripleo-puppet-elements.

Related-Bug: #1954456

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I331e05321601a7436283b4815cf0cd8915a3462b