Remove tftp healthcheck from ironic-pxe image
This is failing downstream because the installed curl-minimal does not
have tftp support. This healthcheck likely doesn't add much value
since dnsmasq is a self-contained tool which will be bound to the port
if it is running.

The healthcheck for dnsmasq as a DHCP server has already been
removed[1].

The alternative to removing this check is to ensure the 'curl' package
is installed instead of 'curl-minimal'.

[1] https://review.opendev.org/c/openstack/tripleo-common/+/799946

Change-Id: I126b997608edb217b0fbf40f4228d0f0af689b0e
(cherry picked from commit 1a73f9561ffacf00ebc3b8def2322c835f225afb)