Add chroot support to tripleo-mount-image
This change adds support for immediately starting a chroot when an
image has been mounted.

This is convenient since many uses of tripleo-mount-image will require
running commands inside the image (such as dnf install). Starting a
useful chroot is slightly involved since it requires bind-mounting
/dev and copying in /etc/resolve.conf. This change handles that, as
well as cleanup and automatic unmount on exit.

This change also makes the -m <directory> command optional, when not
specified a temp directory will be created and used as the mount
directory.

Change-Id: I533ce13833bbfaab0ede3a11b6c22e45fe767939