Move the ceph monitoring stack containers to quay.io
quay.ceph.io is no longer accessible and the Ceph Org moved to
quay.io. This patch just updates the container image references
to point to the new registry.

We are also marking Ussuri content-provider and victoria upgrade jobs
non-voting because Ussuri content provider is failing with similiar
error - Unable to pull containers from quay.ceph.io and its not
possible to fix Ussuri as its EOL and not accepting patches anymore.

[1] https://zuul.opendev.org/t/openstack/build/3997166729c9416b8ed634db844e0302

Closes-Bug: #1973115
Change-Id: I5673ff0480341cce4e0680587648c10c82ab1285