Populate /var/lib/ironic/tftpboot in ironic-pxe image build
This change creates and populates /var/lib/ironic/tftpboot with
required bootloader files.

The same copy operations are done by puppet-ironic[1] when given
tripleo-heat-templates opinionated defaults[2]. This change is
proposed to reduce puppet-ironic's role to only generate configuration
files, and to prepare for a tool other than puppet-ironic (like
ironic-operator) to not be responsible for copying files during init.

The map-file generation is also removed, because we use `dnsmasq`, not
`tftp` to serve tftp content (and the path was wrong anyway).

[1] https://opendev.org/openstack/puppet-ironic/src/branch/master/manifests/pxe.pp
[2] https://opendev.org/openstack/tripleo-heat-templates/src/branch/master/deployment/ironic/ironic-conductor-container-puppet.yaml#L552-L560

Change-Id: I94ad66804a634a4d5880332c2d1637112d654717