Update ovmf-amdsev from ovmf-cc to boot UEFI VMs
Fix for a package was expected in edk2-20220526git16779ede2d36-2.el9,
but it's not there yet.

Without this fix, Nova cannot boot UEFI instances for x86_64 arch because
there is no nvram-template data in
/usr/share/qemu/firmware/edk2-ovmf-amdsev.json.

Symlinking it from /usr/share/qemu/firmware/edk2-ovmf-cc.json would be
quick and dirty w/a but it's better to merge the nvram-template into
it. This way other amdsev-specific feautes would be preserved.

Provide that merging data w/a for nova virtqemud containers build
scripts, until it is fixed in centos-stream9 and rhel9 packaging.

Nova compute also reads those ovmf files, so apply the w/a to
nova compute TCIB as well. It is not clear where from these files
come into nova compute container (edk2-ovmf is only installed in
the libvirt virtqemud container, and no packages require it in
RDO), so for safety reasons, add a quarding check for the source file
presense.

In Wallaby, it should build on cs8 and cs9, so add the test -f guards
for the libvirt container as well.

Change-Id: Ieaacb67d16e82e9c34fcfb31398da5a95cdc3d43
Related: rhbz#2090752
Related: rhbz#2109644
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>
(cherry picked from commit 615cf9ff6ee9cc5ff10ec2b6824862112ee97c03)