Revert "Populate /var/lib/ironic/tftpboot in ironic-pxe image build"
In a podified control plane, /var/lib/ironic/tftpboot needs to be part
of a shared volume used by the conductor and dnsmasq containers. This
means it will be empty regardless of what is on the image, and it
should be populated by the init container for the pod.

This is a modified revert, the original /tftpboot preparation is still
removed.

This reverts commit bcfa3edcee3a48de5b520b22ecf6ff1fb5ad7d4e.

Change-Id: Ic0248cc89b71768afe3dc1e24c42a72b74ea5c3b