Drop collections/roles list installed from ansible-galaxy
requirements.yaml contains the list of collections and roles
of tripleo-ansible. These collections and roles are already
shipped as a dependencies of tripleo-ansible rpm[1]. It does not
make sense to re-install it from galaxy.

We need to keep this file if someone wants to install custom
collections and roles in the tripleo-ansible-ee container.

[1]. https://github.com/rdo-packages/tripleo-ansible-distgit/blob/rpm-master/tripleo-ansible.spec#L25

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ia7ff64c01cb470745f7f9857c7d984f83b36113d