Drop ANSIBLE_RUNNER_IMAGE from tcib_args
In Downstream, Openshift Build System(OSBS) is used to build
the container image.

In tripleo-ansible-ee yaml file, ANSIBLE_RUNNER_IMAGE is overriden
via tcib_args during tcib container build process but
OSBS does not overrides the ANSIBLE_RUNNER_IMAGE arg from FROM
instructions due to
https://github.com/containerbuildsystem/dockerfile-parse/issues/118.

It is not clear when the above issue will get fixed.
In order to unblock the pipeline and to have a proper fix
in place, we are going to using the default image name in tcib_from
argument.

In Downstream, tcib_from will be overriden with downstream
image.

Related-Bug: rhbz#2128230

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I81f911500ba620b27707075ef4fda8d52b2a5831