tripleo-mount-image: replace most lsblk calls with blkid
The version of lsblk on centos8 is missing parameters which this
script relies on. This change switches some lsblk calls to blkid
equivalent to improve support on hosts older than centos9.

There is a similar change proposed to diskimage-builder[1]. This
change addresses a specific failure on a wallaby cs8 job failure[2],
so this change will be backported.

[1] https://review.opendev.org/c/openstack/diskimage-builder/+/850882
[2] https://review.rdoproject.org/zuul/build/3830a8cf87944db495783e43ac5d41bf/log/job-output.txt#5281

(cherry picked from 704cdb267faf67675eec86c28bf633bf29524651)
Change-Id: I03342e8107a214341990584f039dd5fdca2fece4