Fix token cache expires_at check
Tokens can have a expires_at field (specifically, we're seeing this issue
with Satellite). The token does not an expires_in. This patch allows for
the lookup lookup of either key from the token to resolve
cached token authentication issues.

Example token:
```
{'token': '$2a$09$1b6453892473a467d0737uJ56J02CLEXEkOaX0pdLWgLMlg.qWDSi',
 'expires_at': '2022-10-28T04:54:58.296Z',
 'issued_at': '2022-10-17T23:45:09.306Z'}
```

It's unnecessary to get the local time, since tokens have Zulu time. So just
datetime.now() is sufficient. But to perform the comparison, we need to convert
the string datetime from the return value into a valid datetime object.

This change corrects the describe behavior.

Resolves: rhbz#2134075
Change-Id: Ic96679548a553960f1242850e42f9eec6e4a45f0