Support IPv6 for HTTP request
This http-healthcheck always fails in the case the request is against to
IPv6 address.

Usually, the socket in IPv6 return a tuple that contains four values but
only two values in IPv4.

Currently, only two variables are defined to store return values.
Therefore when extract the return value of socket.getpeername() in IPv6,
an ValueError exception is raised and then healthcheck fails.

This change restrict the return value of the socket.getpeername() to
return only first two elements regardless of IPv4/IPv6.

Closes-Bug: #2004268
Change-Id: I57ae54518a86f93ec92f0e74efadf1fa4da29f0f