Inventory: Track seen nets, add/get new nets only
Before this change we fetched many duplicats of the
same network - i.e for every port even if the ports
are on a network we already collected. Subsequently
we got the subnet of all networks in the list which
contained many duplicates.

With this change we keep track of seen network's and
only get networks and subnets from the network service
once.

This reduces the number of calls to neutron by a huge
amount in larger scale deployments. In my 250 node
test enviornment for comparision.

OLD: Generate ansible inventory | localhost | 0:09:35.990553 | 124.56s
NEW: Generate ansible inventory | localhost | 0:07:27.473631 | 2.57s

Change-Id: I5d62c3429d09397cbd9ba0e5a675d2f263e7e031