remove unused tripleo-common dep form nova-scheduler.
Previously tripleo-common was installed in nova-scheduler container so
that the additional scheduler filter for baremetal porisioning can be
used. However nova was already removed from undercloud and the library
is no longer used since wallaby.

Also the package pulls the transitive deps and removing it reduces
the container size significantly.

Change-Id: Icae34a6d3f6e9b2c5a02b3bf52eabba1712ec922