Pin nova_migration user/group for puppet configs
When puppet configures authorized_keys, it applies the nova_migration
group. That group has a dynamic ID in RDO, and the order of packages
installed into config container during its build may wary and impact
the resulted GID for the nova_migration user.

That currently breaks live-migration for mixed-rhel scenarios.

Pin the nova_migration user to maintain the same ID that it used to
have for rhel8 based containers.

Change-Id: I2813d5fc4e12d8b9532cbc32a14cc088328efe62
Related: rhbz#2181897
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>