Remove nomodeset from usage
In investigating packet loss issues, it was discovered that the
root cause to the packet loss issues were the graphical updates
on the console from logging, or other activites such as
interacting with the system console.

In particular, this was discovered to be the case with embedded
graphics controllers which are also in with onboard graphics in
physical servers.

i.e. A server using an aspeed/matrox VGA graphics controller
would be affected, whereas the same server using an add-in GPU
would not experience the same issue.

The root cause, in itself is the use of the ``nomodeset`` option
which is intended more for troubleshooting and provides
framebuffer memory, which results in the NUMA node being locked
upon updates creating the expeienced latency/packet loss
depending on the specific conditions of the machine.

Resolves: rhbz#2179366
Change-Id: I27228ac65000c1a1f17925a3bcc82adffe518c10