Remove audit=1 as it can overflow before the service starts
It appears on some systems and configurations, the default kernel
audit buffer can overflow before the userspace service starts.

Considering this option previously was not passed through, we're
going to remove it for now, and if operators wish to re-enable the
default, they can do so and any issues can be worked on a case
by case basis.

Resolves: rhbz#2190166
Change-Id: I8e00ff1802cbd4e8ad01876e03ca1a59e7541517