Remove outdated kernel options
The nofb kernel option appears to no longer exist, and was
reported non-working to kernel.org in 2016[0]

And it turns out the nousb parameter comes from what is commonly
compiled in as a module, and was incorrectly documented[1]. Thus
the nousb parameter did nothing.

And upon further research, it turns out the vga option is basically
out of date given advancements with grub2, so we are removing it
as well.

Instead of make further changes, just remove the parameters and
enable operators to enable the settings if they so desire.

[0]: https://bugzilla.kernel.org/show_bug.cgi?id=120551
[1]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=40d581489adbb
[2]: https://www.kernel.org/doc/Documentation/svga.txt

Change-Id: Ie764dbca89ada123a39cd80e9a98cc4ee78545b2