Train-only: Make swift internal endpoint detection more robust
Current logic can be confused in case public endpoint (or even admin
endpoint) host contains "internal". This makes the logic rely on
the feature of openstackclient to look up the internal endpoint to
avoid that problem.

This fix is submitted to Train only because swift was removed from
undercloud in Wallaby.

Resolves: rhbz#2190308
Change-Id: I5385309b848bddeaf43fcb660709d1e6540b6c58