Only modify the container manifest if the mediaType has changed.
Only modify the manifest if the mediaType has actually changed.  This is a
partial fix for https://bugzilla.redhat.com/show_bug.cgi?id=2213672 where the
json pretty print with indent=3 causes the SHA256 of the manifest to change
since the manifest is not pretty printed the same way in the source registry.

When the SHA256 changes, the above bug is triggered since the code does
not account for the fact that the httpd type-map file also needs to be
updated to use the changed SHA256.

This will need to be backported to stable/train as well.

Change-Id: I04f6ac171b10af7a294819d6248eac641090cc49
Signed-off-by: James Slagle <jslagle@redhat.com>
(cherry picked from commit b0962d2ba09fbb4da33daa328e6a50cac5e3ba05)