Handle different format in expires_at field
It seems Satellite can return the timestamp in the expires_at field
in a different formant. This updates the logic so that we can support
different formats, especially the one with timezone offset.

Backport note:
stable/train supports pyton 2.7 which does not support the format
string, so the usage is adjusted.

Resolves: rhbz#2229383
Change-Id: I0aa1428f45ad92fd73f3ade9a077f45701a5f4ab
(cherry picked from commit 17c5bda43383726e6caaff4278f7680a0607c233)